Fix depth image fallback dtype and make resolve_asset_path public
- Depth image fallback used np.zeros default float64 instead of the
  spec's float32, causing dtype mismatches in observation buffers.
  Now passes spec.dtype explicitly.
- SceneBuilder._resolve_asset_path was accessed from PyBulletSimulator,
  violating privacy. Renamed to public resolve_asset_path and updated
  all internal callers and tests.

+++ src/surg_rl/rl/observation.py
@@ -580,7 +580,7 @@ class ObservationBuilder:
         elif obs_type == ObservationType.DEPTH_IMAGE:
             if observation.depth_image is not None:
                 return observation.depth_image[:, :, np.newaxis]
-            return np.zeros(spec.shape)
+            return np.zeros(spec.shape, dtype=spec.dtype)
 
         elif obs_type == ObservationType.SEGMENTATION:
             if observation.segmentation is not None:
